rename get to manage
